requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
Django==4.1.2
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4